backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
scikit-learn==1.5.0
xgboost==2.1.0
pandas==2.2.0
numpy==1.26.0
chromadb==0.5.0
sentence-transformers==3.0.0
langchain==0.3.0
langchain-community==0.3.0
langchain-text-splitters==0.3.0
pypdf==4.0.0
httpx==0.27.0
groq==0.11.0
python-multipart==0.0.9
python-dotenv==1.0.1
